mozjpeg::decompress

Struct Decompress

pub struct Decompress<R> { /* private fields */ }

Get pixels out of a JPEG file

High-level wrapper for jpeg_decompress_struct

let d = Decompress::new_path("image.jpg")?;

Implementations

impl Decompress<()>

pub const fn builder() -> DecompressBuilder<'static>

This is DecompressBuilder::new()

impl Decompress<BufReader<File>>

pub fn new_path<P: AsRef<Path>>(path: P) -> Result<Self>

Decode file at path

pub fn new_file(file: File) -> Result<Self>

Decode an already-opened file

impl<'mem> Decompress<&'mem [u8]>

pub fn new_mem(mem: &'mem [u8]) -> Result<Self>

Decode from a JPEG file already in memory

impl<R> Decompress<R>

pub fn new_reader(reader: R) -> Result<Self>
where R: BufRead,

Decode from an io::BufRead, which is BufReader wrapping any io::Read.

pub fn components(&self) -> &[CompInfo]

pub fn color_space(&self) -> COLOR_SPACE

pub fn gamma(&self) -> f64

It’s generally bogus in libjpeg

pub fn pixel_density(&mut self) -> Option<PixelDensity>

Get pixel density of an image from the JFIF APP0 segment. Returns None in case of an invalid density unit.

pub fn markers(&self) -> MarkerIter<'_>

Markers are available only if you enable them via with_markers()

pub fn size(&self) -> (usize, usize)

width,height

pub fn width(&self) -> usize

pub fn height(&self) -> usize

pub fn rgb(self) -> Result<DecompressStarted<R>>

Start decompression with conversion to RGB

pub fn to_colorspace( self, colorspace: ColorSpace, ) -> Result<DecompressStarted<R>>

Start decompression with conversion to colorspace

pub fn rgba(self) -> Result<DecompressStarted<R>>

Start decompression with conversion to RGBA

pub fn grayscale(self) -> Result<DecompressStarted<R>>

Start decompression with conversion to grayscale.

pub fn dct_method(&mut self, method: DctMethod)

Selects the algorithm used for the DCT step.

pub fn do_fancy_upsampling(&mut self, value: bool)

pub fn do_block_smoothing(&mut self, value: bool)

If true, interblock smoothing is applied in early stages of decoding progressive JPEG files; if false, not. Default is true. Early progression stages look “fuzzy” with smoothing, “blocky” without. In any case, block smoothing ceases to be applied after the first few AC coefficients are known to full accuracy, so it is relevant only when using buffered-image mode for progressive images.

pub fn raw(self) -> Result<DecompressStarted<R>>

pub fn image(self) -> Result<Format<R>>

Start decompression without colorspace conversion

pub fn scale(&mut self, numerator: u8)

Rescales the output image by numerator / 8 during decompression. numerator must be between 1 and 16. Thus setting a value of 8 will result in an unscaled image.