mozjpeg

Enum ColorSpace

#[repr(C)]
pub enum ColorSpace {

    JCS_UNKNOWN = 0,
    JCS_GRAYSCALE = 1,
    JCS_RGB = 2,
    JCS_YCbCr = 3,
    JCS_CMYK = 4,
    JCS_YCCK = 5,
    JCS_EXT_RGB = 6,
    JCS_EXT_RGBX = 7,
    JCS_EXT_BGR = 8,
    JCS_EXT_BGRX = 9,
    JCS_EXT_XBGR = 10,
    JCS_EXT_XRGB = 11,
    JCS_EXT_RGBA = 12,
    JCS_EXT_BGRA = 13,
    JCS_EXT_ABGR = 14,
    JCS_EXT_ARGB = 15,
    JCS_RGB565 = 16,

}

Variants

JCS_UNKNOWN = 0

error/unspecified

JCS_GRAYSCALE = 1

monochrome

JCS_RGB = 2

red/green/blue as specified by the RGB_RED, RGB_GREEN, RGB_BLUE, and RGB_PIXELSIZE macros

JCS_YCbCr = 3

Y/Cb/Cr (also known as YUV)

JCS_CMYK = 4

C/M/Y/K

JCS_YCCK = 5

Y/Cb/Cr/K

JCS_EXT_RGB = 6

red/green/blue

JCS_EXT_RGBX = 7

red/green/blue/x When out_color_space it set to JCS_EXT_RGBX, JCS_EXT_BGRX, JCS_EXT_XBGR, or JCS_EXT_XRGB during decompression, the X byte is undefined, and in order to ensure the best performance, libjpeg-turbo can set that byte to whatever value it wishes.

JCS_EXT_BGR = 8

blue/green/red

JCS_EXT_BGRX = 9

blue/green/red/x

JCS_EXT_XBGR = 10

x/blue/green/red

JCS_EXT_XRGB = 11

x/red/green/blue

JCS_EXT_RGBA = 12

Use the following colorspace constants to ensure that the X byte is set to 0xFF, so that it can be interpreted as an opaque alpha channel.

red/green/blue/alpha

JCS_EXT_BGRA = 13

blue/green/red/alpha

JCS_EXT_ABGR = 14

alpha/blue/green/red

JCS_EXT_ARGB = 15

alpha/red/green/blue

JCS_RGB565 = 16

5-bit red/6-bit green/5-bit blue